// src/api/enums/CoreConnectionStatusServiceStatus.ts
// Copyright (c) 2017-2020, The Particl Market developers
// Distributed under the GPL software license, see the accompanying
// file COPYING or https://github.com/particl/particl-market/blob/develop/LICENSE

export enum CoreConnectionStatusServiceStatus {
    ERROR = 'ERROR',
    COOKIESERVICE_DISCONNECTED = 'COOKIESERVICE_DISCONNECTED',
    WALLET_NOT_INITIALIZED = 'WALLET_NOT_INITIALIZED',
    CONNECTED = 'CONNECTED',
    DISCONNECTED =  'DISCONNECTED'
}