// src/api/enums/ObserverStatus.ts
// Copyright (c) 2017-2020, The Particl Market developers
// Distributed under the GPL software license, see the accompanying
// file COPYING or https://github.com/particl/particl-market/blob/develop/LICENSE

export enum ObserverStatus {
    ERROR = 'ERROR',
    RUNNING = 'RUNNING',
    STOPPED = 'STOPPED'
}